win_exec: refactor PS exec runner
