+++ test/integration/targets/win_csharp/aliases
@@ -1,0 +1,2 @@
+shippable/windows/group4
+shippable/windows/smoketest
