+++ test/integration/targets/win_csharp/module_utils/Ansible.Test.cs
@@ -1,0 +1,16 @@
+//AssemblyReference -Name System.DirectoryServices.AccountManagement.dll
+
+using System;
+using System.DirectoryServices.AccountManagement;
+
+namespace Ansible.Test
+{
+    public class Principal
+    {
+        public static string GetCurrentPrincipalSid()
+        {
+            return UserPrincipal.Current.Sid.Value;
+        }
+    }
+}
+
